Version 3.0, Change Crawling Library from HtmlUnit to Jsoup
There are two reasons for this.

First, HtmlUnit is too slow for crawling.
For example, it takes a while to find DOM element using CSS selector and it takes so long to get a web page.

Second, HtmlUnit sometimes crashes in the background task.
I checked the logs but the crash points and the crash causes were different everytime, thus, I still do not know why exactly.
It might be Android limits the resources for background tasks, which is insufficient for HtmlUnit.
By the way, it worked without any problem on desktop and foreground task in Android.

After change to Jsoup, it worked well and the crawling becomes faster than before.
However, the app still uses HtmlUnit for log in, which takes some time to finish.
For now, I think delayed background task problem is solved although I should further examine the logs to confirm.
This is the reason that I set the version to 3.0 not 2.2.