Fix a bug that occurs when there is a board with time condition.
The instructor of the course can create the board(forum) with time condition.
This kind of board does not have link and this induces IllegalArgumentException.
If this bug occurs, then the AsyncTask fails and cannot proceed other stuff and make a notification.
I have fixed this code by just simply adding try catch statement.